requirements.txt
alembic==1.7.5
Babel==2.9.1
black==22.1.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.16.0
email-validator==1.1.3
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
langdetect==1.0.9
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
requests==2.26.0
rq==1.10.1
six==1.16.0
SQLAlchemy==1.4.28
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
wrapt==1.13.3
WTForms==3.0.0
